+++ src/python/dxpy/scripts/dx.py
@@ -3407,6 +3407,12 @@ def watch(args):
     if not re.match("^job-[0-9a-zA-Z]{24}$", args.jobid):
         err_exit(args.jobid + " does not look like a DNAnexus job ID")
 
+    job_describe = dxpy.describe(args.jobid)
+    if 'outputReusedFrom' in job_describe and job_describe['outputReusedFrom'] is not None:
+      args.jobid = job_describe['outputReusedFrom']
+      if not args.quiet:
+        print("Output reused from %s" %(args.jobid))
+
     log_client = DXJobLogStreamClient(args.jobid, input_params=input_params, msg_callback=msg_callback,
                                       msg_output_format=args.format, print_job_info=args.job_info)
 
@@ -5495,9 +5501,9 @@ parser_publish = subparsers.add_parser('publish', help='Publish an app or a glob
                                    prog='dx publish')
 parser_publish.add_argument('executable',
                             help='ID or name and version of an app/global workflow, e.g. myqc/1.0.0').completer = DXPathCompleter(classes=['app', 'globalworkflow'])
-parser_publish.add_argument('--make_default',
-                            help='Set a "default" alias on the published version',
-                            action='store_true', dest='make_default')
+parser_publish.add_argument('--no-default',
+                            help='Do not set a "default" alias on the published version',
+                            action='store_false', dest='make_default')
 parser_publish.set_defaults(func=publish)
 register_parser(parser_publish)